+++ packages/events/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @veloxts/events
 
+## 0.6.102
+
+### Patch Changes
+
+- chore: upgrade fastify to 5.7.4 across all packages
+- Updated dependencies
+  - @veloxts/core@0.6.102
+
 ## 0.6.101
 
 ### Patch Changes